fixed compiler warnings on FreeBSD
persist-state.c: In function `persist_state_load_v4':
persist-state.c:546: warning: long unsigned int format, int arg (arg 3)

cfg-grammar.y: In function `main_parse':
cfg-grammar.y:385: warning: null format string
cfg-grammar.y:393: warning: null format string
cfg-grammar.y:399: warning: null format string

block-ref-grammar.y: In function `block_ref_parse':
block-ref-grammar.y:307: warning: implicit declaration of function `strdup'

pragma-grammar.y: In function `__process_version_string':
pragma-grammar.y:286: warning: implicit declaration of function `strchr'
pragma-grammar.y: In function `pragma_parse':
pragma-grammar.y:374: warning: implicit declaration of function `strdup'

rewrite-expr-grammar.y: In function `rewrite_expr_parse':
rewrite-expr-grammar.y:364: warning: implicit declaration of function `strcmp'

csvparser-grammar.c:2697: warning: implicit declaration of function `csvparser_lex'
csvparser-grammar.y:314: warning: implicit declaration of function `csvparser_error'

+++ src/persist-state.c
@@ -544,7 +544,7 @@ persist_state_load_v4(PersistState *self)
                   if (chain_ofs == 0 || chain_ofs > file_size)
                     {
                       msg_error("Persistent file format error, key block chain offset is too large or zero",
-                                evt_tag_printf("key_block", "%08lx", (gchar *) key_block - (gchar *) map),
+                                evt_tag_printf("key_block", "%08lx", (gulong) ((gchar *) key_block - (gchar *) map)),
                                 evt_tag_printf("key_size", "%d", key_size),
                                 evt_tag_int("ofs", chain_ofs),
                                 NULL);
